voice_assistant/requirements.txt
SpeechRecognition
pyttsx3
pyaudio
wikipedia
colorama
pywin32
AppOpener
pywhatkit
flask
fastapi
uvicorn
google-generativeai
